+++ test/value_at_risk.jl
@@ -39,4 +39,6 @@ end
     returns_big = rand(BigFloat, 16)
     αbig = big(0.05)
     @test @inferred(value_at_risk(returns_big, αbig)) isa BigFloat
+    @test @inferred(value_at_risk(returns_big, αbig; method=:gaussian)) isa BigFloat
+    @test @inferred(value_at_risk(returns_big, αbig; method=:cornish_fisher)) isa BigFloat
 end